+++ cmd/genarate.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/spf13/cobra"
 
-	"github.com/snapp-incubator/crafting-table/internal/generator"
+	"github.com/snapp-incubator/crafting-table/internal/repository"
 	"github.com/snapp-incubator/crafting-table/internal/structure"
 )
 
@@ -16,9 +16,7 @@ var (
 	source      string
 	destination string
 	packageName string
-	getVars     *[]structure.Variables
 	get         string
-	updateVars  *[]structure.UpdateVariables
 	update      string
 	create      bool
 )
@@ -35,12 +33,12 @@ func init() {
 
 	err := generateCMD.MarkFlagRequired("source")
 	if err != nil {
-		panic(err)
+		log.Fatal(err)
 	}
 
 	err = generateCMD.MarkFlagRequired("destination")
 	if err != nil {
-		panic(err)
+		log.Fatal(err)
 	}
 
 	// TODO: add flag for table name
@@ -62,32 +60,34 @@ func generate(_ *cobra.Command, _ []string) {
 	destination = strings.Replace(destination, " ", "", -1)
 
 	if get == "" && update == "" && !create {
-		log.Fatal("You must set at least one flag for get, update or create")
+		log.Fatal("you must set at least one flag for get, update or create")
 	}
 
+	var getVars *[]structure.Variables
 	if get != "" {
 		for strings.Contains(get, " ") {
 			get = strings.Replace(get, " ", "", -1)
 		}
 		get = strings.Replace(get, " ", "", -1)
-		if err := validateFlag(get); err != nil {
+		if err := parser.ValidateGetFlag(get); err != nil {
 			log.Fatal(err)
 		}
-		getVars = parser.ParseVariables(get)
+		getVars = parser.ExtractGetVariables(get)
 	}
 
+	var updateVars *[]structure.UpdateVariables
 	if update != "" {
 		for strings.Contains(update, " ") {
 			update = strings.Replace(update, " ", "", -1)
 		}
 		update = strings.Replace(update, " ", "", -1)
-		if err := validateUpdateFlag(update); err != nil {
+		if err := parser.ValidateUpdateFlag(update); err != nil {
 			log.Fatal(err)
 		}
-		updateVars = parser.ParseUpdateVariables(update)
+		updateVars = parser.ExtractUpdateVariables(update)
 	}
 
-	if err := generator.GenerateRepository(source, destination, packageName, getVars, updateVars, create); err != nil {
+	if err := repository.Generate(source, destination, packageName, getVars, updateVars, create); err != nil {
 		log.Fatal(err)
 	}
 }