+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # UNRELEASED
 
 * Clean up packages. (2022-05-20, @n25a, !15)
+* Issue #13: Renaming variables and packages. (2022-06-17, @n25a, !16)
 
 ## v1.1.1 - May 17 2022